+++ tests/__init__.py
@@ -191,6 +191,7 @@ def tmp_app_with_data(request):
             },
             "creator_username": "queen",
             "frozen_at": 1536238185.881941,
+            "annotations": {"type": "fruit"},
         }
         register_dataset(dataset_info)
 
@@ -211,6 +212,7 @@ def tmp_app_with_data(request):
         },
         "creator_username": "queen",
         "frozen_at": 1536238185.881941,
+        "annotations": {"type": "fruit"},
     }
     register_dataset(dataset_info)