+++ tests/test_dataset_routes.py
@@ -184,6 +184,7 @@ def test_dataset_register_route(tmp_app_with_users):  # NOQA
         },
         "creator_username": "olssont",
         "frozen_at": 1536238185.881941,
+        "annotations": {"software": "bowtie2"},
     }
 
     for token in [dopey_token, sleepy_token]:
@@ -242,6 +243,7 @@ def test_dataset_register_route(tmp_app_with_users):  # NOQA
         },
         "creator_username": "olssont",
         "frozen_at": 1536238185.881941,
+        "annotations": {"software": "bowtie2"},
     }
     r = tmp_app_with_users.post(
         "/dataset/register",
@@ -322,6 +324,7 @@ def test_dataset_register_route_when_created_at_is_string(tmp_app_with_users):  
         "creator_username": "olssont",
         "frozen_at": 1536238185.881941,
         "created_at": "1536238185.881941",
+        "annotations": {"software": "bowtie2"},
     }
 
     headers = dict(Authorization="Bearer " + grumpy_token)